Improve the functionality of the code. used context for all functionlity for my expenses. used "Passing the Entire Expense List to ExpenseList" which is more efficient way.
