+++ src/index.js
@@ -2,10 +2,13 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import "./index.css";
 import App from "./App";
+import { ExpenseProvider } from "./contexts/ExpenseContext";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
   <React.StrictMode>
-    <App />
+    <ExpenseProvider>
+      <App />
+    </ExpenseProvider>
   </React.StrictMode>
 );
