+++ src/pages/Home.jsx
@@ -1,17 +1,5 @@
-const Home = () => {
-  return (
-    <>
-      <div className="container">
-        <div>
-          <img
-            className="w-full"
-            src="https://images.pexels.com/photos/5466785/pexels-photo-5466785.jpeg?auto=compress&cs=tinysrgb&w=800"
-            alt=""
-          />
-        </div>
-      </div>
-    </>
-  );
-};
+import React from "react";
 
-export default Home;
+export default function Home() {
+  return <div>Home Page</div>;
+}
